check celery pid file before rm it

+++ start.sh
@@ -12,7 +12,7 @@ python3 manage.py celery -A FasterRunner.mycelery worker -l info -P solo --setti
 #nohup python3 manage.py celery beat -l info > ./logs/beat.log 2>&1 &
 #nohup python3 manage.py FasterRunner.mycelery beat -l info > ./logs/beat.log 2>&1 &
 #nohup celery -A FasterRunner.mycelery beat -l info > ./logs/beat.log 2>&1 &
-rm celerybeat.pid
+if [ -f celerybeat.pid ]; then rm celerybeat.pid;fi 
 python3 manage.py  celery -A FasterRunner.mycelery beat -l info --settings=FasterRunner.settings.pro --logfile=./logs/beat.log 2>&1 &
 
 # start fastrunner
